feat: 폴더 페이지 링크 없는 경우, 빈 디자인 보이도록 구현
