+++ README.md
@@ -3,7 +3,7 @@
 - [x] 링크 공유 페이지 url path는 ‘/shared’가 되도록 만들어 주세요.
 - [ ] 폴더 페이지의 url path는 ‘/folder’가 되도록 설정해 주세요.
 ## ui
-- [ ] 폴더 페이지에 저장된 링크가 없는 경우 해당(저장된 링크가 없습니다) 디자인이 보이도록 만들어 주세요.
+- [x] 폴더 페이지에 저장된 링크가 없는 경우 해당(저장된 링크가 없습니다) 디자인이 보이도록 만들어 주세요.
 - [x] 폴더 페이지에서 상단 네비게이션 바는 스크롤 시 상단에 고정하지 않고, 가려지도록 해주세요.
 - [ ] 링크 추가하기, 폴더 추가, 공유, 이름 변경, 삭제 버튼들의 기능은 추후 만듭니다.
 - [ ] “전체” 폴더를 선택한 경우 “공유”, “이름 변경”, “삭제” 버튼들이 보이지 않지만, 다른 폴더를 선택한 경우에는 버튼들이 보이도록 해주세요.
