+++ src/Folder.jsx
@@ -1,28 +1,44 @@
+import { useState, useEffect } from 'react';
+
+import { getFolder } from './apis/api';
+
 import SearchBar from './components/SearchBar';
 import Card from './components/Card';
 import './Folder.css';
 
-const Folder = ({items}) => {
-  if(!items || !items.links) {
-    return (
-      <div>파트 1에서 만들었던 랜딩 페이지 적용</div>
-    )
-  }
+const Folder = () => {
+  const [folders, setFolders] = useState({});
+
+  useEffect(() => {
+    const fetchFolder = async () => {
+      //const { folder } = await getFolder();
+      
+      //setFolders(folder);
+    };
+
+    fetchFolder();
+  }, []);
+
+  const isShowFolders = folders && folders.links;
 
   return (
     <div className='container-folder'>
       <SearchBar className='search-bar_ly'/>
-      <ul className='card-frame_ly'> 
-        {items.links.map(({ createdAt, imageSource, title, description, url, id }) => (
-          <Card key={id} 
-            imgSrc={imageSource}
-            title={title}
-            description={description}
-            createdAt={createdAt}
-            url={url}
-          />
-        ))}
-      </ul>
+      {isShowFolders ? (
+        <ul className='card-frame_ly'> 
+          {folders.links.map(({ createdAt, imageSource, title, description, url, id }) => (
+            <Card key={id} 
+              imgSrc={imageSource}
+              title={title}
+              description={description}
+              createdAt={createdAt}
+              url={url}
+            />
+          ))}
+        </ul>
+      ) : (
+        <div className='card-empty'>저장된 링크가 없습니다</div>
+      )}
     </div>
   )
 }
